requirements.txt
pyautogui==0.9.54
PyYAML==6.0.1
langchain==0.3.20
langchain-anthropic==0.3.9